+++ PWGCF/FEMTOSCOPY/FemtoDream/AliFemtoDreamCascadeCuts.h
@@ -26,29 +26,29 @@ class AliFemtoDreamCascadeCuts {
   bool GetIsMonteCarlo() {return fMCData;};
   void SetContributionSplitting(bool contrSplit){
     fContribSplitting=contrSplit;};
-  void SetXiMassRange(double mass,double width){
+  void SetXiMassRange(float mass,float width){
     fcutXiMass=true;fXiMass=mass;fXiMassWidth=width;};
   void SetXiCharge(int charge){fcutXiCharge=true;fXiCharge=charge;}
-  void SetCutXiDaughterDCA(double maxDCA){
+  void SetCutXiDaughterDCA(float maxDCA){
     fcutDCAXiDaug=true;fMaxDCAXiDaug=maxDCA;};
-  void SetCutXiMinDistBachToPrimVtx(double minDist){
+  void SetCutXiMinDistBachToPrimVtx(float minDist){
     fcutMinDistVtxBach=true;fMinDistVtxBach=minDist;};
-  void SetCutXiCPA(double cpa){fcutCPAXi=true;fCPAXi=cpa;};
-  void SetCutXiTransverseRadius(double minRad,double maxRad){
+  void SetCutXiCPA(float cpa){fcutCPAXi=true;fCPAXi=cpa;};
+  void SetCutXiTransverseRadius(float minRad,float maxRad){
     fcutXiTransRadius=true;fMinXiTransRadius=minRad;fMaxXiTransRadius=maxRad;}
-  void Setv0MassRange(double mass,double width){
+  void Setv0MassRange(float mass,float width){
     fcutv0Mass=true;fv0Mass=mass;fv0Width=width;}
-  void SetCutv0MaxDaughterDCA(double maxDCA){
+  void SetCutv0MaxDaughterDCA(float maxDCA){
     fcutv0MaxDCADaug=true;fv0MaxDCADaug=maxDCA;}
-  void SetCutv0CPA(double CPA){
+  void SetCutv0CPA(float CPA){
     fcutCPAv0=true;fCPAv0=CPA;}
-  void SetCutv0TransverseRadius(double minRad,double maxRad){
+  void SetCutv0TransverseRadius(float minRad,float maxRad){
     fcutv0TransRadius=true;fMinv0TransRadius=minRad;fMaxv0TransRadius=maxRad;}
-  void SetCutv0MinDistToPrimVtx(double minDist){
+  void SetCutv0MinDistToPrimVtx(float minDist){
     fcutv0MinDistVtx=true;fv0MinDistVtx=minDist;}
-  void SetCutv0MinDaugDistToPrimVtx(double minDist){
+  void SetCutv0MinDaugDistToPrimVtx(float minDist){
     fcutv0DaugMinDistVtx=true;fv0DaugMinDistVtx=minDist;};
-  void SetRejectOmegas(double mass, double width) {
+  void SetRejectOmegas(float mass, float width) {
     fRejOmega=true;fRejOmegaMass=mass;fRejOmegaWidth=width;
   }
   void SetPDGCodeCasc(int PDG){fPDGCasc=PDG;};
@@ -81,36 +81,36 @@ class AliFemtoDreamCascadeCuts {
   bool fMCData;
   bool fContribSplitting;
   bool fcutXiMass;
-  double fXiMass;
-  double fXiMassWidth;
+  float fXiMass;
+  float fXiMassWidth;
   bool fcutXiCharge;
   int fXiCharge;
   bool fcutDCAXiDaug;
-  double fMaxDCAXiDaug;
+  float fMaxDCAXiDaug;
   bool fcutMinDistVtxBach;
-  double fMinDistVtxBach;
+  float fMinDistVtxBach;
   bool fcutCPAXi;
-  double fCPAXi;
+  float fCPAXi;
   bool fcutXiTransRadius;
-  double fMinXiTransRadius;
-  double fMaxXiTransRadius;
+  float fMinXiTransRadius;
+  float fMaxXiTransRadius;
   bool fcutv0Mass;
-  double fv0Mass;
-  double fv0Width;
+  float fv0Mass;
+  float fv0Width;
   bool fcutv0MaxDCADaug;
-  double fv0MaxDCADaug;
+  float fv0MaxDCADaug;
   bool fcutCPAv0;
-  double fCPAv0;
+  float fCPAv0;
   bool fcutv0TransRadius;
-  double fMinv0TransRadius;
-  double fMaxv0TransRadius;
+  float fMinv0TransRadius;
+  float fMaxv0TransRadius;
   bool fcutv0MinDistVtx;
-  double fv0MinDistVtx;
+  float fv0MinDistVtx;
   bool fcutv0DaugMinDistVtx;
-  double fv0DaugMinDistVtx;
+  float fv0DaugMinDistVtx;
   bool fRejOmega;
-  double fRejOmegaMass;
-  double fRejOmegaWidth;
+  float fRejOmegaMass;
+  float fRejOmegaWidth;
   int fPDGCasc;
   int fPDGv0;
   int fPDGPosDaug;
