+++ PWGCF/FEMTOSCOPY/FemtoDream/AliFemtoDreamZVtxMultContainer.cxx
@@ -59,7 +59,7 @@ void AliFemtoDreamZVtxMultContainer::PairParticlesSE(
     std::vector<std::vector<AliFemtoDreamBasePart>> &Particles,
     AliFemtoDreamCorrHists *ResultsHist,int iMult)
 {
-  double RelativeK = 0;
+  float RelativeK = 0;
   int _intSpec1 = 0;
   int HistCounter=0;
   //First loop over all the different Species
@@ -102,7 +102,7 @@ void AliFemtoDreamZVtxMultContainer::PairParticlesME(
     std::vector<std::vector<AliFemtoDreamBasePart>> &Particles,
     AliFemtoDreamCorrHists *ResultsHist,int iMult)
 {
-  double RelativeK = 0;
+  float RelativeK = 0;
   int _intSpec1 = 0;
   int HistCounter=0;
   auto itPDGPar1 = fPDGParticleSpecies.begin();
@@ -141,7 +141,7 @@ void AliFemtoDreamZVtxMultContainer::PairParticlesME(
     ++itPDGPar1;
   }
 }
-double AliFemtoDreamZVtxMultContainer::RelativePairMomentum(TVector3 Part1Momentum,
+float AliFemtoDreamZVtxMultContainer::RelativePairMomentum(TVector3 Part1Momentum,
                                                             int PDGPart1,
                                                             TVector3 Part2Momentum,
                                                             int PDGPart2)
@@ -149,7 +149,7 @@ double AliFemtoDreamZVtxMultContainer::RelativePairMomentum(TVector3 Part1Moment
     if(PDGPart1 == 0 || PDGPart2== 0){
       AliError("Invalid PDG Code");
     }
-  double results = 0.;
+  float results = 0.;
   TLorentzVector SPtrack,TPProng,trackSum,SPtrackCMS,TPProngCMS;
   //Even if the Daughter tracks were switched up during PID doesn't play a role here cause we are
   //only looking at the mother mass
@@ -159,10 +159,10 @@ double AliFemtoDreamZVtxMultContainer::RelativePairMomentum(TVector3 Part1Moment
                   TDatabasePDG::Instance()->GetParticle(PDGPart2)->Mass());
   trackSum = SPtrack + TPProng;
 
-  double beta = trackSum.Beta();
-  double betax = beta*cos(trackSum.Phi())*sin(trackSum.Theta());
-  double betay = beta*sin(trackSum.Phi())*sin(trackSum.Theta());
-  double betaz = beta*cos(trackSum.Theta());
+  float beta = trackSum.Beta();
+  float betax = beta*cos(trackSum.Phi())*sin(trackSum.Theta());
+  float betay = beta*sin(trackSum.Phi())*sin(trackSum.Theta());
+  float betaz = beta*cos(trackSum.Theta());
 
   SPtrackCMS = SPtrack;
   TPProngCMS = TPProng;
